Include `env` variables defined in a `dependency` in the step itself.
Prior to this commit, a `dependency` defined by the `molecule.yaml` file
would not honor the values in its `env` map. Consider the following:

```yaml
dependency:
  name: shell
  command: env
  env:
    FOO: bar
```

Invoking `molecule test` would show that `FOO` is not set during the run
of this dependency. The cause was the fact that these values were never
actually passed to the `util.run_command()` invocation. This also means
`MOLECULE_EPHEMERAL_DIRECTORY` and similar `MOLECULE_*` environment
variables that are defined by the dependency's `default_env` property
were also never available to the dependency's environment.