Change linkchecker in tox.ini to run against docs dir
Change linkchecker to run against docs dir to validate only links
provided by user instead of all auto-generated links from site dir.
All auto-generated links in site directory contains absolute links.
The problem is that linkchecker tries to traverse those links and
is getting 404 error, which is expected as pages are not yet published.
This blocks pull-request and adding new pages or restructuring existing
docsite as in this pr[1].

[1]https://github.com/ansible/molecule/pull/4044