Adding back molecule template based on driver.name
In past when using cookicuter was possible to have templates based on a
set of variables.
This commit adds a default image entry when using 'containers',
'podman' or 'docker' drivers.
The `init-scenario.yml` also will by default only copy the create.yml
and destroy.yml if not using one of those container drivers.