Fix test method name. (#4165)
This appears to have been missed while reviewing #4135. If I understand
the [pytest
docs](https://docs.pytest.org/en/8.0.x/explanation/goodpractices.html#test-discovery)
correctly, the method is not recognized as a test without the change.

This pull request is nearly identical to #4164, which I accidentally
closed and was unable to reopen.