Remove non-functional html report (#4528)
This remove some HTML reporting functionality based on rich and the
console. With stdout/stderr changes to the logger and the the
interception of stdout/stderr for command borders this functionality
would produce empty files.

Searching sourceforge and google, there is little to no mention of it or
apperant use.

It was not tested nor exposed at the CLI.